requirements.txt
boto3==1.15.1
celery==4.4.7
Django==2.2.15
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-minio-storage==0.3.8
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.11.1
filetype==1.0.7
flake8==3.8.3
Pillow==7.2.0
psycopg2>=2.7.5,<=2.8.5
pyotp==2.4.0
python-decouple==3.3
redis==3.5.3
stripe==2.51.0
twilio==6.45.4
